UPDATED: preferences screen
This commit goes with updated preferences screen.
