+++ DarkModeDiscovery/DarkModeDiscovery/DemoContent/Preferences/PreferencesVC.swift
@@ -1,5 +1,5 @@
 //
-//  PreferencesViewController.swift
+//  PreferencesVC.swift
 //  DarkModeDiscovery
 //
 //  Created by Mikhail Zhigulin in 7531.
@@ -14,6 +14,12 @@ import Cocoa
 
 class PreferencesViewController: NSViewController {
 
+    @IBOutlet weak var segmentedControl: NSSegmentedControl!
+
+    @IBAction func segmentedControlValueChanged(_ sender: NSSegmentedCell) {
+        changeDarkModeValue(selected: sender.selectedSegment)
+    }
+
     override func viewDidLoad() {
         super.viewDidLoad()
 
@@ -25,5 +31,30 @@ class PreferencesViewController: NSViewController {
         super.viewDidAppear()
 
         self.parent?.view.window?.title = self.title!
+        updateDarkModeOption()
+    }
+
+    private func updateDarkModeOption() {
+        switch AppearanceService.DarkModeUserChoice {
+        case .auto:
+            segmentedControl.selectedSegment = 2
+        case .on:
+            segmentedControl.selectedSegment = 1
+        case .off:
+            segmentedControl.selectedSegment = 0
+        }
+    }
+
+    private func changeDarkModeValue(selected: Int) {
+        switch selected {
+        case 0:
+            changeDarkModeManually(.off)
+        case 1:
+            changeDarkModeManually(.on)
+        case 2:
+            changeDarkModeManually(.auto)
+        default:
+            break
+        }
     }
 }
